+++ src/model/polling-response.ts
@@ -1,0 +1,13 @@
+import {PollingResponseError} from "./polling-response-error";
+
+/**
+ * Model of polling response
+ *
+ * @author Christopher Holomek
+ * @since 06.01.2020
+ */
+export interface PollingResponse {
+    jsonrpc: string,
+    result?: any[],
+    error?: PollingResponseError,
+}